+++ src/main/java/sol/shinhansecuirty/sosolbe/DTO/StockInfoDTO.java
@@ -1,4 +1,4 @@
-package sol.shinhansecuirty.sosolbe.Option.DTO;
+package sol.shinhansecuirty.sosolbe.DTO;
 
 import lombok.Builder;
 import lombok.Getter;
