+++ src/main/java/sol/shinhansecuirty/sosolbe/Service/OptionService.java
@@ -1,7 +1,7 @@
-package sol.shinhansecuirty.sosolbe.Option.Service;
+package sol.shinhansecuirty.sosolbe.Service;
 
 import org.springframework.stereotype.Service;
-import sol.shinhansecuirty.sosolbe.Option.DTO.StockInfoDTO;
+import sol.shinhansecuirty.sosolbe.DTO.StockInfoDTO;
 
 import java.util.ArrayList;
 import java.util.Arrays;
